Add Track Request functionality
This commit adds an additional function which will generate a file with
all the charmhub tracks that need to be available for a certain release.
It will not yet check with charmhub if these tracks already exist.

It also will upload this file as an artifact when the github action is
executed.