add procedure and script to split up a review of our issue backlog
This is some simple instructions and a script to take out entire issue backlog and split it up, by repository, into groups of repositories with roughly the same number of issues.  This makes it easy for us to then each take a group of a few repos, review the open issues, and maintain our backlog.

See readme.md for more details