fix: Prefix artifact name with job name and run id (#153)
Since v4 of the upload-artifact, artifacts are immutable. Thus, 
enable optional input `artifact-prefix` in order to avoid uploading logs
from different jobs to the same artifact. This input is neccesary when
the action is being called multiple times for the same job using a
matrix. On top of that, always prefix the artifact with the job name.

Fix #152